+++ lib/open_api_spex/json_api_helpers.ex
@@ -1,69 +1,22 @@
 defmodule OpenApiSpex.JsonApiHelpers do
   alias OpenApiSpex.JsonApiHelpers.{JsonApiDocument, JsonApiResource}
-  alias OpenApiSpex.Schema
 
-  def document_schema(%JsonApiDocument{} = document) do
-    if not is_binary(document.title) do
-      raise "%JsonApiDocument{} :title is required and must be a string"
-    end
-
-    resource = document.resource
-    resource_item_schema = resource_schema(resource)
-
-    resource_schema =
-      if document.multiple do
-        %Schema{type: :array, items: resource_item_schema, title: resource.title <> "List"}
-      else
-        resource_item_schema
-      end
-
-    %Schema{
-      type: :object,
-      properties: %{
-        data: resource_schema
-      },
-      required: [:data],
-      title: document.title <> "Document"
-    }
+  def document_schema(document) do
+    JsonApiDocument.schema(document)
   end
 
-  def resource_schema(%JsonApiResource{} = resource) do
-    if not is_binary(resource.title) do
-      raise "%JsonApiResource{} :title is required and must be a string"
-    end
-
-    %Schema{
-      type: :object,
-      properties: %{
-        id: %Schema{type: :string},
-        type: %Schema{type: :string},
-        attributes: attributes_schema(resource)
-      },
-      required: [:id, :type],
-      title: resource.title <> "Resource"
-    }
-  end
-
-  def attributes_schema(%JsonApiResource{} = resource) do
-    if not is_binary(resource.title) do
-      raise "%JsonApiResource{} :title is required and must be a string"
-    end
-
-    %Schema{
-      type: :object,
-      properties: resource.properties,
-      title: resource.title <> "Attributes"
-    }
+  def resource_schema(resource) do
+    JsonApiResource.schema(resource)
   end
 
   defmacro generate_document_schema(attrs) do
     quote do
       require OpenApiSpex
 
-      @document struct!(OpenApiSpex.JsonApiHelpers.JsonApiDocument, unquote(attrs))
+      @document struct!(JsonApiDocument, unquote(attrs))
       def document, do: @document
 
-      @document_schema OpenApiSpex.JsonApiHelpers.document_schema(@document)
+      @document_schema JsonApiDocument.schema(@document)
       def document_schema, do: @document_schema
 
       OpenApiSpex.schema(@document_schema)
@@ -74,10 +27,10 @@ defmodule OpenApiSpex.JsonApiHelpers do
     quote do
       require OpenApiSpex
 
-      @resource struct!(OpenApiSpex.JsonApiHelpers.JsonApiResource, unquote(attrs))
+      @resource struct!(JsonApiResource, unquote(attrs))
       def resource, do: @resource
 
-      @resource_schema OpenApiSpex.JsonApiHelpers.resource_schema(@resource)
+      @resource_schema JsonApiResource.schema(@resource)
       def resource_schema, do: @resource_schema
 
       OpenApiSpex.schema(@resource_schema)
